added eroor functionality for charts
the html controlling the error display is in each page that displays a chart while the css is in global.css. Meanwhile, the JS is achieved by adding all functionalities that are used to display the chart into a function. This function has a try block for handling display of charts, and catch block for handling error in displaying chart.